# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Banking System API",
        "description": "Full API Documentation for the Banking System",
        "version": "1.0.0"
    },
    "paths": {
        "/api/test": {
            "get": {
                "tags": [
                    "Test"
                ],
                "summary": "Test Endpoint to verify Swagger",
                "operationId": "51d98c8ce3052b28283bde47c9996025",
                "responses": {
                    "200": {
                        "description": "Swagger is working!"
                    }
                }
            }
        },
        "/api/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get all user accounts",
                "operationId": "e2c6417368ae098c813b932c8a7726a8",
                "responses": {
                    "200": {
                        "description": "List of accounts"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Create a new account",
                "operationId": "325487f288176d8fefe93c68d36202e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "type",
                                    "account_name"
                                ],
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "account_name": {
                                        "type": "string"
                                    },
                                    "balance": {
                                        "type": "number"
                                    },
                                    "interest_rate": {
                                        "type": "number"
                                    },
                                    "credit_limit": {
                                        "type": "number"
                                    },
                                    "minimum_balance": {
                                        "type": "number"
                                    },
                                    "parent_account_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get a single account",
                "operationId": "1705fe31beef661ca7097ab2698431e6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The account data"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Update an account",
                "operationId": "44248934a314e599ac892e42de726cbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "type": {
                                        "type": "string"
                                    },
                                    "account_name": {
                                        "type": "string"
                                    },
                                    "balance": {
                                        "type": "number"
                                    },
                                    "interest_rate": {
                                        "type": "number"
                                    },
                                    "credit_limit": {
                                        "type": "number"
                                    },
                                    "minimum_balance": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Delete an account",
                "operationId": "bf7df2da543d456f53e5fe1a9ca26822",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/accounts/{id}/total-balance": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get total balance including children accounts",
                "operationId": "f090f4a231be921b3b6cd429f75e14d7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Total account balance"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/verify-code": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Verify email using code",
                "operationId": "30951c1eae4688fba36236f81ed0ba0b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "code"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Email verified"
                    },
                    "422": {
                        "description": "Invalid code"
                    }
                }
            }
        },
        "/api/verify-code/resend": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Resend verification code",
                "operationId": "18e435ceeaaf90272c8b8cfac7efe8da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Code resent"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login with email and password",
                "operationId": "e064e69806fcb96ad24fb5a23a62937b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "c79711b0216966b71cabbdca4786cc23",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new customer",
                "operationId": "e4e1cbbe63e77831efe7d6dfcea0eb02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created, verification code sent"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Test",
            "description": "Test"
        },
        {
            "name": "Accounts",
            "description": "Accounts"
        },
        {
            "name": "Auth",
            "description": "Auth"
        }
    ]
}